Support settings files on Python 2
Previously, settings files would be silently ignored on Python 2.x unless the optional PyPI configparser package was installed.

This commits tries the newer import and falls back to the legacy 2.x stdlib name. Additionally, it no longer silently hides errors when a config file exists but is not readable for some reason as the user should know when the config they intended to use cannot be loaded.