Reorder CHANGELOG to be in reverse chronological order
It is a bit unusual and unconventional to order a CHANGELOG with the
most recent release at the bottom. Due to this, it is easy for a new
user to incorrectly read version 4.0.0 as the latest release. Bring
CHANGELOG into more typical ordering to reduce confusion.

Follows suggestions from "Keep a CHANGELOG": http://keepachangelog.com/.

> List releases in reverse-chronological order (newest on top).