Make isort run with Python 3 only
As isort is primarily used as a command-line tool, it is in a good
position to move to Python 3 only. Other command-line-only tools have
begun moving to Python 3. For example: black, mypy, pylint, and Sphinx.
It will continue to support formatting Python 2 code, so users can still
format their Python 2 projects.

Allows for many code cleanups. By removing these workarounds, the
project will be easier to maintain and reduce testing resources.

Python 2 is scheduled to be EOL on January 1, 2020. The Python community
is moving on.

https://www.python.org/dev/peps/pep-0373/

For reasons to drop Python 2 support, see:

https://python3statement.org/

For an overview of projects that have dropped Python 2 support entirely,
see:

https://hugovk.github.io/drop-python/2.7/

Fixes #808