Auto-generate the complete list of stdlib modules
The new script, mkstdlibs.py, uses intersphinx to dump a complete list
of Python stdlib modules to the files py3.txt/py27.txt. These files are
read by isort to decide if a module should be considered stdlib or not.

In the future, if the stdlib changes, the fix is to simply re-run the
script.

Fixes #985