Add dependencies to enable better detection of third-party packages
When used from within pre-commit, isort is installed into a separate
environment that does not have access to anything installed with the
package being developed. However, with these packages installed, we
can use isort's analysis of the package metainformation to identify
which packages are third-party.

Ideally we could specify this using something like
`isort[pipfile,pyproject,requirements]` but I don't think there's a
way to do this using pre-commit.