Clarify documentation for --skip option
The --skip option only handles full file paths or names of
individual path components.

For example, "foo/bar/baz.py" can be used to skip just that file.
And "foo" would skip any files with that name and any files nested
in directories with that name.

On the other hand, "foo/bar" does *not* skip everything in the
"foo/bar" directory. --skip-glob can be used to achieve this.

Attempt to clarify the documentation to say that.