Do not cache configuration files
Fixes #1994

isort can run in long processes when used in an editor.  In that
scenario it is confusing for users when editing a configuration file
does not change isorts outcome.

Remove the caches in `settings.py` completely to re-read the
configuration on each run.

This is safe for the CLI case as we here either use the `Trie`
implementation which collects all known configuration files beforehand,
or we exactly construct one `Config` object based on the first filename,
the working dir, or the explicit `settings_path|file` settings.

Not caching is therefore more efficient for the CLI case and more
useful for the editor case.