CI: use `pip` over `pipx for poetry install
`pipx` is installed on all the runners by default, but using this means
`pipx` is run with the system Python, and not the one installed with
`steup-python`. This was noticed when e.g. the MacOS Python 3.9 job
would report:

    creating virtual environment...
    creating shared libraries...
    upgrading shared libraries...
    installing poetry...
    done! ✨ 🌟 ✨
      installed package poetry 1.3.1, installed using Python 3.13.0
      These apps are now globally available
        - poetry
    Poetry (version 1.3.1)

Python 3.13.0 is the system version pre-installed on these runners[1],
and a similar pattern was seen on the Ubuntu and Windows runners. An
alternative would be to add an install step for `pipx` but this feels
simpler

Link: https://github.com/actions/runner-images/blob/de16eefce8361c24c716958843d8c87cb1c25990/images/macos/macos-14-Readme.md [1]